requirements.txt
duckduckgo-search
langchain-community==0.2.6
openai==1.35.5
streamlit==1.36.0
Pillow==9.5.0
octoai-sdk==0.7.0
